requirements.txt
slack-bolt>=1.20.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-mcp-adapters>=0.1.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0
chromadb>=0.5.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.3.0
pypdf>=4.0.0
hubspot-api-client>=9.0.0
